+++ src/algorithm/boundary_condition/periodic_bc.hpp
@@ -3,7 +3,7 @@
 #include "../BootesArray.hpp"
 
 
-void periodic_boundary_condition_x1i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void periodic_boundary_condition_x1i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                               int &x2s, int &x2l, int &ng2,
                                                               int &x3s, int &x3l, int &ng3){
     #pragma omp parallel for collapse(4)
@@ -18,7 +18,7 @@ void periodic_boundary_condition_x1i(BootesArray<float> &quan, int &x1s, int &x1
     }
 }
 
-void periodic_boundary_condition_x1o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void periodic_boundary_condition_x1o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                               int &x2s, int &x2l, int &ng2,
                                                               int &x3s, int &x3l, int &ng3){
     #pragma omp parallel for collapse(4)
@@ -34,7 +34,7 @@ void periodic_boundary_condition_x1o(BootesArray<float> &quan, int &x1s, int &x1
 }
 
 
-void periodic_boundary_condition_x2i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void periodic_boundary_condition_x2i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                               int &x2s, int &x2l, int &ng2,
                                                               int &x3s, int &x3l, int &ng3){
     #pragma omp parallel for collapse(4)
@@ -49,7 +49,7 @@ void periodic_boundary_condition_x2i(BootesArray<float> &quan, int &x1s, int &x1
     }
 }
 
-void periodic_boundary_condition_x2o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void periodic_boundary_condition_x2o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                               int &x2s, int &x2l, int &ng2,
                                                               int &x3s, int &x3l, int &ng3){
     #pragma omp parallel for collapse(4)
@@ -64,7 +64,7 @@ void periodic_boundary_condition_x2o(BootesArray<float> &quan, int &x1s, int &x1
     }
 }
 
-void periodic_boundary_condition_x3i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void periodic_boundary_condition_x3i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                               int &x2s, int &x2l, int &ng2,
                                                               int &x3s, int &x3l, int &ng3){
     #pragma omp parallel for collapse(4)
@@ -80,7 +80,7 @@ void periodic_boundary_condition_x3i(BootesArray<float> &quan, int &x1s, int &x1
 }
 
 
-void periodic_boundary_condition_x3o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void periodic_boundary_condition_x3o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                               int &x2s, int &x2l, int &ng2,
                                                               int &x3s, int &x3l, int &ng3){
     #pragma omp parallel for collapse(4)
