+++ src/algorithm/boundary_condition/reflective_bc.hpp
@@ -3,7 +3,7 @@
 #include "../BootesArray.hpp"
 
 
-void reflective_boundary_condition_x1i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void reflective_boundary_condition_x1i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                                 int &x2s, int &x2l, int &ng2,
                                                                 int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -24,7 +24,7 @@ void reflective_boundary_condition_x1i(BootesArray<float> &quan, int &x1s, int &
 }
 
 
-void reflective_boundary_condition_x1o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void reflective_boundary_condition_x1o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                                 int &x2s, int &x2l, int &ng2,
                                                                 int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -45,7 +45,7 @@ void reflective_boundary_condition_x1o(BootesArray<float> &quan, int &x1s, int &
 }
 
 
-void reflective_boundary_condition_x2i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void reflective_boundary_condition_x2i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                                 int &x2s, int &x2l, int &ng2,
                                                                 int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -66,7 +66,7 @@ void reflective_boundary_condition_x2i(BootesArray<float> &quan, int &x1s, int &
 }
 
 
-void reflective_boundary_condition_x2o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void reflective_boundary_condition_x2o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                                 int &x2s, int &x2l, int &ng2,
                                                                 int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -87,7 +87,7 @@ void reflective_boundary_condition_x2o(BootesArray<float> &quan, int &x1s, int &
 }
 
 
-void reflective_boundary_condition_x3i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void reflective_boundary_condition_x3i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                                 int &x2s, int &x2l, int &ng2,
                                                                 int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -108,7 +108,7 @@ void reflective_boundary_condition_x3i(BootesArray<float> &quan, int &x1s, int &
 }
 
 
-void reflective_boundary_condition_x3o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void reflective_boundary_condition_x3o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                                 int &x2s, int &x2l, int &ng2,
                                                                 int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
