+++ src/algorithm/hydro/donercell.hpp
@@ -5,23 +5,23 @@
 #include <cmath>
 
 
-void donercell(float &rhoL, float &pL, float &uL,
-          float &rhoR, float &pR, float &uR,
-          BootesArray<float> &valsL,
-          BootesArray<float> &valsR,
-          BootesArray<float> &fluxs,
+void donercell(double &rhoL, double &pL, double &uL,
+          double &rhoR, double &pR, double &uR,
+          BootesArray<double> &valsL,
+          BootesArray<double> &valsR,
+          BootesArray<double> &fluxs,
           int MOMPRESIND,
-          float &gamma){
+          double &gamma){
 
     // step 1: wave speed estimates, Toro 10.5.1
-    float aL = soundspeed(rhoL, pL, gamma);
-    float aR = soundspeed(rhoR, pR, gamma);
-    float sL = uL - aL;
-    float sR = uR + aR;
+    double aL = soundspeed(rhoL, pL, gamma);
+    double aR = soundspeed(rhoR, pR, gamma);
+    double sL = uL - aL;
+    double sR = uR + aR;
     // step 2: calculate flux
     for (int val_ind = 0; val_ind < fluxs.shape()[0]; val_ind ++){
-        float flux_L = valsL(val_ind) * uL;
-        float flux_R = valsR(val_ind) * uR;
+        double flux_L = valsL(val_ind) * uL;
+        double flux_R = valsR(val_ind) * uR;
         if (val_ind == MOMPRESIND){
             flux_L += pL;
             flux_R += pR;
